Move MaternKarhunenLoeveKernel and MaternFeatureMapKernel to separate files and remove the old hyperbolic-only MaternIntegratedKernel, making geometric_kernels.kernels.geometric_kernels module obsolete
