+++ geometric_kernels/kernels/__init__.py
@@ -3,11 +3,8 @@ Geometric kernel baseclass and specific implementations for spaces.
 """
 # noqa: F401
 from geometric_kernels.kernels.base import BaseGeometricKernel
-from geometric_kernels.kernels.geometric_kernels import (
-    MaternFeatureMapKernel,
-    MaternIntegratedKernel,
-    MaternKarhunenLoeveKernel,
-)
+from geometric_kernels.kernels.feature_map import MaternFeatureMapKernel
+from geometric_kernels.kernels.karhunen_loeve import MaternKarhunenLoeveKernel
 from geometric_kernels.kernels.matern_kernel import (
     MaternGeometricKernel,
     default_feature_map,
